made a tensor generator and applier
